+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 ### Added
+- added meter dashboard to myems-api and myems-web
 ### Changed
 - changed carbon unit to KGCO2E in myems-api
 ### Fixed